Refactor tf scaffold
